requirements.txt
fastapi
motor
pydantic
uvicorn
python-dotenv
httpx
psycopg2-binary
requests
sqlparse
sphinx 
sphinx-autodoc-typehints 
sphinx-rtd-theme 
myst-parser